Module dioxus_router::components

Components interacting with the router.

Structs

HistoryButtonProps

The properties for a GoBackButton or a GoForwardButton.

LinkProps

The properties for a Link.

RouterConfigFactory

The config for Router.

RouterProps

The props for Router.

Enums

IntoRoutable

Something that can be converted into a NavigationTarget.

Functions

FailureExternalNavigation

The default component to render when an external navigation fails.

GoBackButton

A button to go back through the navigation history. Similar to a browsers back button.

GoForwardButton

A button to go forward through the navigation history. Similar to a browsers forward button.

Link

A link to navigate to another route.

Outlet

An outlet for the current content.

Router

A component that renders the current route.