Function dioxus_router::hooks::use_navigator

pub fn use_navigator(cx: &ScopeState) -> &Navigator

A hook that provides access to the navigator to change the router history. Unlike [use_router], this hook will not cause a rerender when the current route changes

The Routable macro will define a version of this hook with an explicit type.

#[derive(Clone, Routable)]
enum Route {
    #[route("/")]
    Index {},
    #[route("/:id")]
    Dynamic { id: usize },
}

#[component]
fn App(cx: Scope) -> Element {
    render! {
        Router::<Route> {}
    }
}

#[component]
fn Index(cx: Scope) -> Element {
    let navigator = use_navigator(&cx);

    render! {
        button {
            onclick: move |_| { navigator.push(Route::Dynamic { id: 1234 }); },
            "Go to /1234"
        }
    }
}

#[component]
fn Dynamic(cx: Scope, id: usize) -> Element {
    render! {
        p {
            "Current ID: {id}"
        }
    }
}