Module dioxus_router::routable

Routable

Structs

RouteParseError

An error that occurs when parsing a route.

SiteMapSegment

A type erased map of the site structure.

Enums

SegmentType

The type of a route segment.

Traits

FromQuery

Something that can be created from a query string.

FromRouteSegment

Something that can be created from a route segment.

FromRouteSegments

Something that can be created from route segments.

Routable

Something that can be:

ToRouteSegments

Something that can be converted to route segments.