Trait dioxus_router::routable::FromQuery

pub trait FromQuery {
    // Required method
    fn from_query(query: &str) -> Self;
}

Something that can be created from a query string.

This trait needs to be implemented if you want to turn a query string into a struct.

A working example can be found in the examples folder in the root package under query_segments_demo.

Required Methods

fn from_query(query: &str) -> Self

Create an instance of Self from a query string.