Trait dioxus_router::routable::Routable

pub trait Routable: FromStr + Display + Clone + 'static {
    const SITE_MAP: &'static [SiteMapSegment];

    // Required method
    fn render<'a>(&self, cx: &'a ScopeState, level: usize) -> Element<'a>;

    // Provided methods
    fn is_child_of(&self, other: &Self) -> bool { ... }
    fn parent(&self) -> Option<Self> { ... }
    fn flatten_site_map<'a>(
    ) -> FlatMap<Iter<'a, SiteMapSegment>, Vec<Vec<SegmentType>>, fn(_: &SiteMapSegment) -> Vec<Vec<SegmentType>>> { ... }
    fn static_routes() -> Vec<Self> { ... }
}

Something that can be:

Converted from a route.

Converted to a route.

Rendered as a component.

This trait can be derived using the #[derive(Routable)] macro.

Required Associated Constants

const SITE_MAP: &'static [SiteMapSegment]

The error that can occur when parsing a route.

Required Methods

fn render<'a>(&self, cx: &'a ScopeState, level: usize) -> Element<'a>

Render the route at the given level

Provided Methods

fn is_child_of(&self, other: &Self) -> bool

Checks if this route is a child of the given route.

Example

use dioxus_router::prelude::*;
use dioxus::prelude::*;

#[component]
fn Home(cx: Scope) -> Element { todo!() }
#[component]
fn About(cx: Scope) -> Element { todo!() }

#[derive(Routable, Clone, PartialEq, Debug)]
enum Route {
    #[route("/")]
    Home {},
    #[route("/about")]
    About {},
}

let route = Route::About {};
let parent = Route::Home {};
assert!(route.is_child_of(&parent));

fn parent(&self) -> Option<Self>

Get the parent route of this route.

Example

use dioxus_router::prelude::*;
use dioxus::prelude::*;

#[component]
fn Home(cx: Scope) -> Element { todo!() }
#[component]
fn About(cx: Scope) -> Element { todo!() }

#[derive(Routable, Clone, PartialEq, Debug)]
enum Route {
    #[route("/home")]
    Home {},
    #[route("/home/about")]
    About {},
}

let route = Route::About {};
let parent = route.parent().unwrap();
assert_eq!(parent, Route::Home {});

fn flatten_site_map<'a>( ) -> FlatMap<Iter<'a, SiteMapSegment>, Vec<Vec<SegmentType>>, fn(_: &SiteMapSegment) -> Vec<Vec<SegmentType>>>

Returns a flattened version of Self::SITE_MAP.

fn static_routes() -> Vec<Self>

Gets a list of all the static routes. Example static route: #[route("/static/route")]