Function dioxus_router::hooks::use_route

pub fn use_route<R: Routable + Clone>(cx: &ScopeState) -> Option<R>

A hook that provides access to information about the current routing location.

The Routable macro will define a version of this hook with an explicit type.

Return values

None, when not called inside a Link component.

Otherwise the current route.

Panic

When the calling component is not nested within a Link component during a debug build.

Example

#[derive(Clone, Routable)]
enum Route {
    #[route("/")]
    Index {},
}

#[component]
fn App(cx: Scope) -> Element {
    render! {
        h1 { "App" }
        Router::<Route> {}
    }
}

#[component]
fn Index(cx: Scope) -> Element {
    let path: Route = use_route(&cx).unwrap();
    render! {
        h2 { "Current Path" }
        p { "{path}" }
    }
}