Struct dioxus_router::prelude::WebHistory

pub struct WebHistory<R: Routable> { /* private fields */ }

A HistoryProvider that integrates with a browser via the History API.

Prefix

This HistoryProvider supports a prefix, which can be used for web apps that aren’t located at the root of their domain.

Application developers are responsible for ensuring that right after the prefix comes a /. If that is not the case, this HistoryProvider will replace the first character after the prefix with one.

Application developers are responsible for not rendering the router if the prefix is not present in the URL. Otherwise, if a router navigation is triggered, the prefix will be added.

Implementations

impl<R: Routable> WebHistory<R>

pub fn new(prefix: Option<String>, do_scroll_restoration: bool) -> Selfwhere <R as FromStr>::Err: Display,

Create a new WebHistory.

If do_scroll_restoration is true, WebHistory will take control of the history state. It’ll also set the browsers scroll restoration to manual.